@@ -1,41 +1,45 @@
-'use strict';
+import { readdir } from 'node:fs';
+import { join } from 'node:path';
 
-const fs = require('fs');
-const path = require('path');
-
-const doxray = require('doxray');
+import doxray from 'doxray';
 
 // relative path from the root directory of remedocs
 // to the root directory of any package to be documented
 const remeDir = './';
 
-// relative path from remeDir to njk documents inside the project
-const includeDir = path.join(remeDir, 'content/_includes/');
+// relative path from remeDir to filters inside the project
+const filterDir = join(remeDir, 'src/filters/');
 
 // define the documentation regex…
 const doxOptions = {
   regex: {
-    njk: {
-      opening: /\{#*\s*@docs[^\n]*\n/m,
-      closing: /#}/,
-      comment: /\{#\s*@docs\s*([^#]|#[^}])*\s*#}/gm,
-      ignore: /\{#\s*@no-docs[\s\S]*/gm,
+    js: {
+      opening: /\/\*\s*@docs[^\n]*\n/m,
+      closing: /\*\//,
+      comment: /\/\*\s*@docs[^*]*\*+(?:[^/*][^*]*\*+)*\//gm,
+      ignore: /\/\*\s*ignore-@docs[\s\S]*/gm,
+    },
+    cjs: {
+      opening: /\/\*\s*@docs[^\n]*\n/m,
+      closing: /\*\//,
+      comment: /\/\*\s*@docs[^*]*\*+(?:[^/*][^*]*\*+)*\//gm,
+      ignore: /\/\*\s*ignore-@docs[\s\S]*/gm,
     },
   },
 };
 
-module.exports = () =>
+export default () =>
   new Promise((resolve, reject) => {
     // get the docs
-    fs.readdir(includeDir, (err, files) => {
+    readdir(filterDir, (err, files) => {
       if (err) {
         reject(err);
       }
 
-      const njk = files
-        .filter((file) => file.endsWith('.macros.njk'))
+      const js = files
+        .filter((file) => file.endsWith('.js') || file.endsWith('.cjs'))
         .map((file) => {
-          const filePath = path.join(includeDir, file);
+          const filePath = join(filterDir, file);
           const docs = doxray([filePath], doxOptions);
           const info = docs.patterns.find(
             (pattern) => String(pattern.category).toLowerCase() === 'file',
@@ -53,7 +57,6 @@ module.exports = () =>
             ),
           };
         });
-
-      resolve(njk);
+      resolve(js);
     });
   });
