@@ -1,25 +1,23 @@
-/* eslint-disable no-process-env, no-sync */
+/* eslint-disable no-process-env */
 
-'use strict';
+import rss from '@11ty/eleventy-plugin-rss';
+import syntaxHighlight from '@11ty/eleventy-plugin-syntaxhighlight';
+import fs from 'fs-extra';
+import yaml from 'js-yaml';
+import { concat, groupBy, merge } from 'lodash-es';
 
-const rss = require('@11ty/eleventy-plugin-rss');
-const syntaxHighlight = require('@11ty/eleventy-plugin-syntaxhighlight');
-const fs = require('fs-extra');
-const yaml = require('js-yaml');
-const _ = require('lodash');
+import * as birds from '#filters/birds.js';
+import * as events from '#filters/events.js';
+import * as images from '#filters/images.js';
+import * as mentions from '#filters/mentions.js';
+import * as pages from '#filters/pages.js';
+import * as tags from '#filters/tags.js';
+import * as taxonomy from '#filters/taxonomy.js';
+import * as time from '#filters/time.js';
+import * as type from '#filters/type.js';
+import * as utils from '#filters/utils.js';
 
-const birds = require('#/birds.cjs');
-const events = require('#/events.cjs');
-const images = require('#/images.cjs');
-const mentions = require('#/mentions.cjs');
-const pages = require('#/pages.cjs');
-const tags = require('#/tags.cjs');
-const taxonomy = require('#/taxonomy.cjs');
-const time = require('#/time.cjs');
-const type = require('#/type.cjs');
-const utils = require('#/utils.cjs');
-
-module.exports = (eleventyConfig) => {
+export default (eleventyConfig) => {
   eleventyConfig.setUseGitIgnore(false);
   eleventyConfig.setWatchThrottleWaitTime(100);
   eleventyConfig.addPlugin(rss);
@@ -89,9 +87,9 @@ module.exports = (eleventyConfig) => {
   );
 
   // filters
-  eleventyConfig.addFilter('concat', _.concat);
-  eleventyConfig.addFilter('merge', _.merge);
-  eleventyConfig.addFilter('group', _.groupBy);
+  eleventyConfig.addFilter('concat', concat);
+  eleventyConfig.addFilter('merge', merge);
+  eleventyConfig.addFilter('group', groupBy);
 
   eleventyConfig.addFilter('typeCheck', utils.typeCheck);
   eleventyConfig.addFilter('styles', utils.styles);
@@ -181,6 +179,7 @@ module.exports = (eleventyConfig) => {
     eleventyConfig.on('eleventy.after', () => {
       if (process.env.IMAGE_CACHE_CHANGED) {
         // If the image cache has been updated, emit the new JSON file
+        // eslint-disable-next-line no-sync
         fs.outputJsonSync(images.CACHE_FILE, images.imageCache, { spaces: 2 });
       }
     });
