@@ -1,12 +1,16 @@
 /* eslint-disable no-sync */
 
-const eleventyImg = require('@11ty/eleventy-img');
+import { jest } from '@jest/globals';
 
-const { image } = require('#/images');
+jest.unstable_mockModule('@11ty/eleventy-img', () => ({
+  default: jest.fn(),
+}));
 
-jest.mock('@11ty/eleventy-img');
+const eleventyImg = await import('@11ty/eleventy-img');
+const { image } = await import('#filters/images');
 
-eleventyImg.statsSync.mockReturnValue({
+eleventyImg.default.generateHTML = jest.fn();
+eleventyImg.default.statsSync = jest.fn().mockReturnValue({
   jpeg: [{ url: '/assets/images/img-960w.webp' }],
 });
 
@@ -27,19 +31,19 @@ describe('image filters', () => {
     test('calls eleventy-img plugin with options', () => {
       image(src, 'alt text', { class: 'foobar' });
 
-      expect(eleventyImg).toHaveBeenCalledTimes(1);
-      expect(eleventyImg.mock.calls[0][0]).toEqual(src);
+      expect(eleventyImg.default).toHaveBeenCalledTimes(1);
+      expect(eleventyImg.default.mock.calls[0][0]).toEqual(src);
 
-      const options = eleventyImg.mock.calls[0][1];
+      const options = eleventyImg.default.mock.calls[0][1];
 
       expect(options.widths).toEqual([480, 960, 1600]);
       expect(options.formats).toEqual(['webp', 'jpeg']);
       expect(options.filenameFormat('hash', src, 480, 'webp')).toBe(
         'img-480w.webp',
       );
-      expect(eleventyImg.statsSync).toHaveBeenCalledTimes(1);
-      expect(eleventyImg.generateHTML).toHaveBeenCalledTimes(1);
-      expect(eleventyImg.generateHTML.mock.calls[0][1]).toEqual({
+      expect(eleventyImg.default.statsSync).toHaveBeenCalledTimes(1);
+      expect(eleventyImg.default.generateHTML).toHaveBeenCalledTimes(1);
+      expect(eleventyImg.default.generateHTML.mock.calls[0][1]).toEqual({
         alt: 'alt text',
         sizes: '(min-width: 45em) 50vw, 100vw',
         loading: 'lazy',
@@ -51,8 +55,8 @@ describe('image filters', () => {
     test('can override sizes', () => {
       image(src, null, null, 'gallery');
 
-      expect(eleventyImg.generateHTML).toHaveBeenCalledTimes(1);
-      expect(eleventyImg.generateHTML.mock.calls[0][1]).toEqual({
+      expect(eleventyImg.default.generateHTML).toHaveBeenCalledTimes(1);
+      expect(eleventyImg.default.generateHTML.mock.calls[0][1]).toEqual({
         alt: '',
         sizes: '(min-width: 95em) 30vw, 50vw',
         loading: 'lazy',
